+++ stocal/examples/events.py
@@ -16,6 +16,6 @@ process = stocal.Process([
 
 
 if __name__ == '__main__':
-    traj = process.trajectory({}, tmax=100)
+    traj = process.sample({}, tmax=100)
     for _ in traj:
         print(traj.time, traj.state['A'], traj.state['A2'])